+++ src/app/app.module.ts
@@ -22,14 +22,12 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatIconModule } from '@angular/material/icon';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
-
+import { MarsComponent } from './pages/mars/mars.component';
 import { MoonComponent } from './pages/moon/moon.component';
-import { MarsComponent } from './components/mars/mars.component';
 
 import { SocketComponent } from './socket/socket.component';
 
 
-import { CountdownComponent } from './components/countdown/countdown.component';
 import { NavbarComponent } from './components/navbar/navbar.component';
 
 
@@ -43,11 +41,9 @@ import { NavbarComponent } from './components/navbar/navbar.component';
     SigninComponent,
     MoonComponent,
     MarsComponent,
-
     SocketComponent,
 
-    NavbarComponent,
-    CountdownComponent,
+    NavbarComponent
 
 
   ],
