--- src/app/components/countdown/countdown.component.html
@@ -1,1 +1,0 @@
-<countdown class="countdown" #cd [config]="config" (event)="handleEvent($event)"></countdown>
