--- src/app/components/countdown/countdown.component.scss
@@ -1,6 +1,0 @@
-.countdown {
-  width: 300px;
-  font-family: 'Lilita One', cursive;
-  font-size: 100px;
-  color: red;
-}
