--- src/app/components/mars/mars.component.html
@@ -1,4 +1,0 @@
-<section>
-    <img src="../../../assets/elf.png" alt="" id="elf1">
-    <img src="../../../assets/elf.png" alt="" id="elf2">
-</section>
