--- src/app/components/mars/mars.component.scss
@@ -1,21 +1,0 @@
-section{
-    background: url('../../../assets/marsbackground.jpg');
-    width: 100%;
-    height: 100%;
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-img{
-    width: 15px;
-    position: relative;
-}
-
-#elf1{
-    top: 8%;
-    left: 92%;
-}
-
-#elf2{
-    top: 18%;
-    left: 32%;
-}
