--- src/app/components/mars/mars.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-mars',
-  templateUrl: './mars.component.html',
-  styleUrls: ['./mars.component.scss']
-})
-export class MarsComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
