+++ src/app/pages/game-homepage/game-homepage.component.html
@@ -36,7 +36,9 @@
         <img src="../../../assets/mercure.png" alt="" class="planet" id="planet2">
     </div>
     <div class="planet-container" id="container3">
-        <img src="../../../assets/mars.png" alt="" class="planet" id="planet3">
+        <a [routerLink]="['/mars']" routerLinkActive="router-link-active" >
+            <img src="../../../assets/mars.png" alt="" class="planet" id="planet3">
+        </a>
     </div>
     <div class="planet-container" id="container4">
         <img src="../../../assets/venus.png" alt="" class="planet" id="planet4">
