+++ src/app/pages/mars/mars.component.html
@@ -1,6 +1,6 @@
 <app-navbar #navbar [title]="title" [score]="score" ></app-navbar>
 <div class="planetImage">
-    <img src="https://media.rainpos.com/7774/moon_colony.jpg" alt="">
+    <img src="https://imgix.bustle.com/uploads/shutterstock/2019/11/8/87804c70-e0a5-4620-ab1e-0d9f4256c515-shutterstock-724077286.jpg?w=1020&h=576&fit=crop&crop=faces&auto=format&q=70" alt="">
   </div>
 
 <div *ngFor="let elf of elfs; let i = index" class="{{elf.name}}">
@@ -12,4 +12,3 @@
   <p> Well Done! </p>
 </div>
 </button>
-
