+++ src/app/pages/mars/mars.component.scss
@@ -8,32 +8,32 @@
 
 .et1 img{
     position: absolute;
-    top: 71.2%;
-    left: 21.8%;
+    top: 82.2%;
+    left: 72.5%;
 }
 
 .elf1 img {
     position: absolute;
-    top: 68%;
-    left: 15.8%; 
+    top: 69%;
+    left: 62.8%; 
 }
 
 .elf2 img {
     position: absolute;
-    top: 75%;
-    left: 48%; 
+    top: 72%;
+    left: 20%; 
 }
 
 .elf3 img {
     position: absolute;
     top: 33%;
-    left: 40%;    
+    left: 20%;    
 }
 
 .elf4 img {
     position: absolute;
-    top: 53.5%;
-    left: 90%;    
+    top: 45.5%;
+    left: 51%;    
 }
 
 .victory {
